# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Transaction {
  id: string;
  account: string;
  type: string;
  /** @nullable */
  credit: number | null;
  /** @nullable */
  debit: number | null;
  posted: string;
  date: string;
}

export interface TransferInput {
  /**
     * @maximum 190939.9
     * @exclusiveMinimum 0
     */
  amount: number;
}

export type ApprovalStatus = typeof ApprovalStatus[keyof typeof ApprovalStatus];


export const ApprovalStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface Approval {
  id: string;
  title: string;
  detail: string;
  status: ApprovalStatus;
}

export type ApprovalUpdateStatus = typeof ApprovalUpdateStatus[keyof typeof ApprovalUpdateStatus];


export const ApprovalUpdateStatus = {
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface ApprovalUpdate {
  status: ApprovalUpdateStatus;
}

export interface DashboardState {
  transfers: Transaction[];
  approvals: Approval[];
}

export type InsightChatMessageRole = typeof InsightChatMessageRole[keyof typeof InsightChatMessageRole];


export const InsightChatMessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface InsightChatMessage {
  role: InsightChatMessageRole;
  /**
     * @minLength 1
     * @maxLength 2000
     */
  content: string;
}

export type InsightChatInputUser = typeof InsightChatInputUser[keyof typeof InsightChatInputUser];


export const InsightChatInputUser = {
  ben: 'ben',
  james: 'james',
} as const;

export type InsightChatInputInsightKey = typeof InsightChatInputInsightKey[keyof typeof InsightChatInputInsightKey];


export const InsightChatInputInsightKey = {
  shortfall: 'shortfall',
  flagged: 'flagged',
  vendor: 'vendor',
} as const;

export interface InsightChatInput {
  user: InsightChatInputUser;
  insightKey: InsightChatInputInsightKey;
  /**
     * @minLength 1
     * @maxLength 1200
     */
  message: string;
  /** @maxItems 8 */
  history: InsightChatMessage[];
}

export interface InsightChatResponse {
  reply: string;
}