# lib/api-zod/src/generated/types/dashboardState.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Approval } from './approval';
import type { Transaction } from './transaction';

export interface DashboardState {
  transfers: Transaction[];
  approvals: Approval[];
}