# lib/api-zod/src/generated/types/insightChatInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { InsightChatInputInsightKey } from './insightChatInputInsightKey';
import type { InsightChatInputUser } from './insightChatInputUser';
import type { InsightChatMessage } from './insightChatMessage';

export interface InsightChatInput {
  user: InsightChatInputUser;
  insightKey: InsightChatInputInsightKey;
  /**
     * @minLength 1
     * @maxLength 1200
     */
  message: string;
  /** @maxItems 8 */
  history: InsightChatMessage[];
}