# lib/api-zod/src/generated/types/insightChatMessage.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { InsightChatMessageRole } from './insightChatMessageRole';

export interface InsightChatMessage {
  role: InsightChatMessageRole;
  /**
     * @minLength 1
     * @maxLength 2000
     */
  content: string;
}