# lib/api-zod/src/generated/types/transferInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TransferInput {
  /**
     * @maximum 190939.9
     * @exclusiveMinimum 0
     */
  amount: number;
}